hyprwater: Fresnel from the real surface; window physics as a slider
Fresnel was a static ring: edgeProximity^2 * strength, knowing nothing
about the water. But a bright rim is not a decoration - water reflects
only ~2% of light head-on and climbs toward 1 at grazing angles, which
is exactly why a droplet on a table is clear in the middle and shows a
bright arc where its surface curves away from you.

So both sources of steepness now feed one slope and one Schlick term:
the meniscus (the window's edge curving down to the desktop, which is
what edgeProximity always described) and the water itself (wave flanks,
and above all the bow wave of a dragged window - the steepest water on
screen). The rim brightens as waves roll into it instead of sitting
there, steep water glints wherever it is, and calm water collapses back
to exactly the old static rim. The ^5 is a gate, not a curve: gentle
swell is ignored, so this picks out features instead of washing the
window out. The slope is free - it is the same gradient the warp is
already built from.

window_physics goes from a toggle back to a 0..1 slider (INT -> FLOAT).
The coupling a toggle could not express is emergent rather than
multiplied in: the slider scales the bow wave, and the bow wave is what
the glass derives its glint from, so turning the physics down fades the
drag highlight with it and zero leaves only the static droplet rim. 1.0
reproduces the old feel exactly, so an existing window_physics = 1 in
anyone's config keeps meaning what it meant. hyprwater-get.sh moves the
key out of INT_KEYS to match the new type.

+++ hypr/hyprwater/src/PluginConfig.cpp
@@ -52,7 +52,7 @@ void registerConfig(HANDLE handle) {
     // either disturb the water or they don't, and the force they were shipped
     // with is the one that looked right. The mouse keeps a slider because how
     // heavy a fingertip feels IS the thing people will disagree about.
-    addConfigValue<Config::Values::Int>(handle, ConfigKeys::SHIMMER_WINDOW_PHYSICS, Config::INTEGER{1});
+    addConfigValue<Config::Values::Float>(handle, ConfigKeys::SHIMMER_WINDOW_PHYSICS, Config::FLOAT{1.0f});
     addConfigValue<Config::Values::Float>(handle, ConfigKeys::SHIMMER_MOUSE, Config::FLOAT{0.3f});
     addConfigValue<Config::Values::Int>(handle, ConfigKeys::SHIMMER_CURRENTS, Config::INTEGER{1});
     // Settings exposes this as a slider snapped to powers of two; any value in
@@ -188,7 +188,7 @@ void initConfigPointers(HANDLE handle, SPluginConfig& config) {
     config.shimmerAgitation        = getStaticPtr<Hyprlang::FLOAT>(handle, ConfigKeys::SHIMMER_AGITATION);
     config.shimmerViscosity        = getStaticPtr<Hyprlang::FLOAT>(handle, ConfigKeys::SHIMMER_VISCOSITY);
     config.shimmerMurk             = getStaticPtr<Hyprlang::FLOAT>(handle, ConfigKeys::SHIMMER_MURK);
-    config.shimmerWindowPhysics    = getStaticPtr<Hyprlang::INT>(handle, ConfigKeys::SHIMMER_WINDOW_PHYSICS);
+    config.shimmerWindowPhysics    = getStaticPtr<Hyprlang::FLOAT>(handle, ConfigKeys::SHIMMER_WINDOW_PHYSICS);
     config.shimmerMouse            = getStaticPtr<Hyprlang::FLOAT>(handle, ConfigKeys::SHIMMER_MOUSE);
     config.shimmerCurrents         = getStaticPtr<Hyprlang::INT>(handle, ConfigKeys::SHIMMER_CURRENTS);
     config.shimmerCurrentsRes      = getStaticPtr<Hyprlang::INT>(handle, ConfigKeys::SHIMMER_CURRENTS_RES);
